+++ Table/commands.py
@@ -9,8 +9,12 @@ def addTable(table):
         begin("Add table %s" % (table))
         cur.execute("create table %s (%s_id serial primary key, %s_modtime timestamp with time zone, %s_author varchar(80))" % (table, table, table, table))
         commit("SUCCESS: CREATE TABLE %s;" % (table))
+        msg = 'def addTable(): passed'
+        return msg
     except:
         rollback("FAILED: CREATE TABLE %s;" % (table))
+        msg = 'def addTable(): failed'
+        return msg
 
 
 # drop table
@@ -25,8 +29,12 @@ def dropTable(table):
             begin("Drop table %s" % (table))
             cur.execute("drop table %s" % (table))
             commit("SUCCESS: DROP TABLE %s;" % (table))
+            msg = 'def dropTable(): passed'
+            return msg
         except:
             rollback("FAILED: DROP TABLE %s;" % (table))
+            msg = 'def dropTable(): failed'
+            return msg
 
 
 # analyse specific table
@@ -45,8 +53,12 @@ def analyzeTable(table):
             raise AssertionError('FAILED: %s does not exist' % (table))
 
         commit('SUCCESS: ANALYZE %s;\n%s' % (table, result))
+        msg = 'def analyzeTable(): passed'
+        return msg
     except:
         rollback('FAILED: ANALYZE %s;' % (table))
+        msg = 'def analyzeTable(): failed'
+        return msg
 
 
 # vacuum specific table
@@ -65,5 +77,40 @@ def vacuumTable(table):
             raise AssertionError('FAILED: %s does not exist' % (table))
 
         commit('SUCCESS: VACUUM FULL VERBOSE ANALYZE %s;\n%s' % (table, result))
+        msg = 'def vacuumTable(): passed'
+        return msg
     except:
         rollback('FAILED: VACUUM FULL VERBOSE ANALYZE %s;' % (table))
+        msg = 'def vacuumTable(): failed'
+        return msg
+
+
+# function for creating a diagram
+def createDia(user, table):
+
+    database = "'%s'" % (databaseInputFields[0].get())
+
+    raise ValueError("Not supported yet.")
+    dumpname = database + ".sql"
+    '''
+    pg_dump test -U raphael > test.sql
+    cat test.sql | pf_dump2graph -n test1 -d svg
+    '''
+    try:
+        databasedump = subprocess.Popen(
+            "pg_dump -s -d %s -U %s > %s" % (database, user, dumpname),
+            shell = True
+        )
+        writeTarget("creating dump successful")
+    except: # describe to concrete exception
+        writeTarget("Error in creating dump")
+    time.sleep(2) # take a short nap so that the dump can be created
+
+    try:
+        subprocess.call(
+            "cat %s | pf_dump2graph -n ./%s -d svg" % (dumpname, database),
+            shell = True
+        )
+        writeTarget("creating svg file successfull")
+    except:
+        writeTarget("creating svg file not successfull")